Extract mapper functions
