+++ livedata/src/main/java/io/getstream/chat/android/livedata/request/QueryChannelsPaginationRequest.kt
@@ -1,9 +1,5 @@
 package io.getstream.chat.android.livedata.request
 
-import io.getstream.chat.android.client.api.models.QueryChannelsRequest
-import io.getstream.chat.android.client.api.models.QuerySort
-import io.getstream.chat.android.client.utils.FilterObject
-
 /**
  * Paginate query channels on the queryChannels repo
  * Similar to QueryChannelsRequest but without the watch, filter and sort params
@@ -15,26 +11,6 @@ data class QueryChannelsPaginationRequest(
     val messageLimit: Int = 10
 ) {
 
-    fun isFirstPage(): Boolean {
-        return channelOffset == 0
-    }
-
-    internal fun toAnyChannelPaginationRequest(): AnyChannelPaginationRequest {
-        return AnyChannelPaginationRequest().apply {
-            this.channelLimit = channelLimit
-            this.channelOffset = channelOffset
-            this.messageLimit = messageLimit
-        }
-    }
-
-    fun toQueryChannelsRequest(filter: FilterObject, sort: QuerySort?, userPresence: Boolean): QueryChannelsRequest {
-        val querySort = sort ?: QuerySort()
-        var request = QueryChannelsRequest(filter, channelOffset, channelLimit, querySort)
-
-        request = request.withMessages(messageLimit)
-        if (userPresence) {
-            request = request.withPresence()
-        }
-        return request.withWatch()
-    }
+    val isFirstPage: Boolean
+        get() = channelOffset == 0
 }
